Refactor state machine code: simplify structure and improve type handling
