+++ Sources/StateMachine/StateMachine.swift
@@ -1,7 +1,5 @@
 import ComposableArchitecture
 
-// MARK: - StateMachine
-
 public protocol StateMachine {
     associatedtype State
     associatedtype Action
@@ -16,8 +14,6 @@ public protocol StateMachine {
     func runIOEffect(_ ioEffect: IOEffect) async -> IOResult?
 }
 
-// MARK: - Transition helpers only
-
 public extension StateMachine {
     static var undefined: Transition { (nil, nil) }
     static var identity: Transition { (nil, nil) }
@@ -25,32 +21,4 @@ public extension StateMachine {
     static func run(_ effect: IOEffect) -> Transition { (nil, effect) }
     static func transition(_ state: State, effect: IOEffect) -> Transition { (state, effect) }
     static func unsafe(_ action: @escaping () -> Void) -> Transition { action(); return (nil, nil) }
-}
-
-// MARK: - Specialized reducer implementation for MappableAction-based Actions
-
-public extension StateMachine where Self: Reducer, Action: MappableAction, Action.Input == Input, Action.IOResult == IOResult {
-    public func reduce(into state: inout State, action: Action) -> Effect<Action> {
-        guard let mapped = Action.map(action) else { return .none }
-        let transition: Transition
-        switch mapped {
-        case .first(let input):
-            transition = Self.reduceInput(state, input)
-        case .second(let ioResult):
-            transition = Self.reduceIOResult(state, ioResult)
-        }
-        return apply(transition, to: &state)
-    }
-
-    @inlinable
-    public func apply(_ transition: Transition, to state: inout State) -> Effect<Action> {
-        if let next = transition.0 { state = next }
-        guard let effect = transition.1 else { return .none }
-
-        return .run { send in
-            if let result = await self.runIOEffect(effect) {
-                await send(Action.ioResult(result))
-            }
-        }
-    }
 }
