+++ Sources/StateMachine/XOR.swift
@@ -1,0 +1,4 @@
+public enum XOR<A, B>: Sendable {
+    case first(A)
+    case second(B)
+}
